Remove Node::id() requirement
A `Node` instance should not need to know about its id, because this
would restrict us in case of IPFS, where we know the ID only _after_
adding the data to the storage.

This commit thus removes the requirement `Node::id()`.

The interface `DagBackend::get()` now returns the `NodeId` for the
`Node` as well. This is mostly because it makes the implementation of
`AsyncDag` less complicated.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Tested-by: Matthias Beyer <mail@beyermatthias.de>